+++ mod.ts
@@ -33,3 +33,9 @@ export {
 } from "./polymorphic.ts";
 /** types */
 export type { InitializerFunction, ResolverFunction } from "./polymorphic.ts";
+
+export { getNew } from "./strategy/utils.ts";
+export type {
+  FunctionSerializable,
+  NewSerializable,
+} from "./strategy/utils.ts";