+++ mod_test.ts
@@ -18,7 +18,8 @@ test({
     assertEquals(typeof mod.polymorphicClassFromJSON, "function");
     assertEquals(typeof mod.PolymorphicResolver, "function");
     assertEquals(typeof mod.PolymorphicSwitch, "function");
+    assertEquals(typeof mod.getNew, "function");
 
-    assertEquals(Object.keys(mod).length, 11);
+    assertEquals(Object.keys(mod).length, 12);
   },
 });