+++ strategy/_utils.ts
@@ -1,8 +1,7 @@
 // Copyright 2018-2021 Gamebridge.ai authors. All rights reserved. MIT license.
-import { Serializable } from "../serializable.ts";
 
-type NewSerializable<T> = T & (new () => Serializable);
-type FunctionSerializable = () => Serializable;
+import { FunctionSerializable, NewSerializable } from "./utils.ts";
+
 /** for strategies */
 export type StrategyTypeArgument<T> = NewSerializable<T> | FunctionSerializable;
 
@@ -12,16 +11,9 @@ export function isObject(obj: any): obj is Record<string, any> {
 }
 
 /** for strategy type arguments */
-function isNewable<T>(
+export function isNewable<T>(
   type: NewSerializable<T> | FunctionSerializable,
 ): type is NewSerializable<T> {
   return (type && typeof type === "function" && type.prototype &&
     type.prototype.constructor) === type;
-}
-
-/** get new strategy type arguments */
-export function getNew<T>(
-  type: NewSerializable<T> | FunctionSerializable,
-): Serializable {
-  return isNewable(type) ? new type() : type();
 }