+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - docs for @SerializeProperty short cut functions
-
+- new `strategy/utils.ts` exposing `getNew(type)` for custom decorators
 
 ## [v1.1.4] - 2021-01-19